Move swap input refs to hook
